post endpoint to add strat

+++ backend/src/main/java/no/edvardsen/backend/models/Lineup.java
@@ -1,26 +1,17 @@
 package no.edvardsen.backend.models;
 
 import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Set;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 /**
@@ -52,16 +43,16 @@ public class Lineup {
     private NadeType nade;
 
     @Column(name = "throw_coordinate_x")
-    private int throwCoordinateX;
+    private double throwCoordinateX;
 
     @Column(name = "throw_coordinate_y")
-    private int throwCoordinateY;
+    private double throwCoordinateY;
 
     @Column(name = "land_coordinate_x")
-    private int landCoordinateX;
+    private double landCoordinateX;
 
     @Column(name = "land_coordinate_y")
-    private int landCoordinateY;
+    private double landCoordinateY;
 
     @Column(name = "video_path")
     private String videoPath; // Path to video showing how to throw the nade
@@ -73,7 +64,7 @@ public class Lineup {
     public Lineup() {
     }
 
-    public Lineup(String name, String desc, NadeType nade, int throwX, int throwY, int landX, int landY,
+    public Lineup(String name, String desc, NadeType nade, double throwX, double throwY, double landX, double landY,
             String videoPath) {
         this.name = name;
         this.desc = desc;
